// src/app/models/fitbod.ts
export class Fitbod {
    Date: string = "";
    "Distance(m)": string = "";
    "Duration(s)": string = "";
    Exercise: string = "";
    Incline: string = "";
    Note: string = "";
    Reps: string = "";
    Resistance: string = "";
    "Weight(kg)": string = "";
    isWarmup: string = "";
    multiplier: string = "";
}